pyproject.toml
[project]
name = "mail-management"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-ninja>=1.4.1",
    "django-ninja-jwt>=5.3.7",
    "mypy>=1.15.0",
    "ruff>=0.11.10",
]
